Move operations on sample_data to it's own module
